src/ckanext-clearml/dev-requirements.txt
pytest-ckan
clearml